#28 feat : row component 서버/클라이언트 분리작업 완료 및 스타일컴포넌트 분리작업 완료, 타입 분리작업 완료
